[IMP] Memory footprint, speed and disk space
Generating and storing a 70MB auditfile takes 8 minutes for qweb, but for
lxml.etree to reconstruct using a namespace takes 3 hours and 1GB of extra
memory.

Further optimization is achieved by sticking to the Odoo prefetch size when
chunking record sets.

As XML compresses very well, the auditfile is now stored in the Odoo database
in compressed format for easier handling and saving disk space.