[IMP] l10n_nl_postcodeapi: Don't do trial query
The service is paid with a limited number of requests per month. We
don't want to burn through the allotted number of queries twice as
fast. We also don't want to give annoying warnings when the only
reason we failed was that we ran out of queries.

This does unfortunately mean that problems are harder to diagnose.