[FIX] l10n_nl_postcodeapi: Don't use ormcache
In the case of get_provider_obj, the method isn't actually
expensive. get_param() has its own cache, and constructing an Api
object is very lightweight.

In the case of get_province(), ormcache explicitly says you shouldn't
return a recordset, and the cost of the database lookup is negligible
compared to the cost of the preceding API call.